+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 -   Update apiary/openapi docs
+-   Update golemio errors ([core#62](https://gitlab.com/operator-ict/golemio/code/modules/core/-/issues/62))
 
 ## [3.5.3] - 2023-05-17